backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
idna==3.10
pillow==11.3.0
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.2.0
whitenoise==6.8.2

